\name{jrb_heatmaps-package}
\alias{jrb_heatmaps-package}
\alias{jrb_heatmaps}
\docType{package}
\title{
\packageTitle{jrb_heatmaps}
}
\description{
\packageDescription{jrb_heatmaps}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jrb_heatmaps}
\packageIndices{jrb_heatmaps}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{jrb_heatmaps}

Maintainer: \packageMaintainer{jrb_heatmaps}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
